CODE: command level connection exposed via Command.connect
The command level connection is the ability to connect to GoodData
using credentials loaded from ~/.gooddata or by issuing an interactive
prompt of the ~/.gooddata file is not present or empty.

The command level connection functionality was available only as an
instance method of a command inherited from GoodData::Command::Base
(and depending on the implementation of the "auth:connect" command).